Revert "Change Custody Count to Uint8 (#14386)" (#14415)
This reverts commit bd7ec3fa976c49cbe67c16f9f20cb8cfcaf882b2.

+++ beacon-chain/p2p/custody.go
@@ -86,7 +86,7 @@ func (s *Service) CustodyCountFromRemotePeer(pid peer.ID) uint64 {
 	if metadata != nil {
 		custodyCount := metadata.CustodySubnetCount()
 		if custodyCount > 0 {
-			return uint64(custodyCount)
+			return custodyCount
 		}
 	}
